backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langgraph>=0.3.0
chromadb>=1.0.0
openai>=1.60.0
python-dotenv>=1.0.0
